# api/openapi.yaml
openapi: 3.1.0
info:
  title: Audiobook TTS Reader Local API
  version: 1.0.0
  description: Local loopback API for controlling Windows SAPI audiobook playback. Mutating requests require the per-run API token printed by `tts-reader serve`.
servers:
  - url: http://127.0.0.1:8080
tags:
  - name: Voices
    description: Installed local Windows SAPI voices.
  - name: Books
    description: Local text files registered for the current API process.
  - name: Playback
    description: Audiobook playback lifecycle and progress controls.
  - name: Events
    description: OpenAPI document and Server-Sent Events stream.
paths:
  /api/openapi.yaml:
    get:
      tags: [Events]
      operationId: getOpenAPISpec
      summary: Get this OpenAPI contract
      security: []
      responses:
        "200":
          description: OpenAPI YAML document
          content:
            application/yaml:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Error"

  /api/v1/voices:
    get:
      tags: [Voices]
      operationId: listVoices
      summary: List installed Windows SAPI voices
      security: []
      responses:
        "200":
          description: Installed voices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoicesResponse"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

  /api/v1/books:
    post:
      tags: [Books]
      operationId: addBook
      summary: Register a local text book
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBookRequest"
      responses:
        "201":
          description: Registered book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"
    get:
      tags: [Books]
      operationId: listBooks
      summary: List registered books for the current process
      security: []
      responses:
        "200":
          description: Registered books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooksResponse"
        "403":
          $ref: "#/components/responses/Error"

  /api/v1/playback:
    post:
      tags: [Playback]
      operationId: startPlayback
      summary: Start audiobook playback
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartPlaybackRequest"
      responses:
        "202":
          description: Playback accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"
    get:
      tags: [Playback]
      operationId: getPlaybackState
      summary: Get current playback state
      security: []
      responses:
        "200":
          description: Current playback state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "403":
          $ref: "#/components/responses/Error"

  /api/v1/playback/pause:
    post:
      tags: [Playback]
      operationId: pausePlayback
      summary: Pause playback between chunks
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      responses:
        "200":
          description: Paused state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "409":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"

  /api/v1/playback/resume:
    post:
      tags: [Playback]
      operationId: resumePlayback
      summary: Resume paused playback
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      responses:
        "200":
          description: Playing state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "409":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"

  /api/v1/playback/stop:
    post:
      tags: [Playback]
      operationId: stopPlayback
      summary: Stop playback and persist the current byte position
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      responses:
        "200":
          description: Stopped state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "409":
          description: Playback is still stopping because the old TTS session did not finish before the request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          description: Engine stop or progress persistence returned an internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/playback/position:
    put:
      tags: [Playback]
      operationId: setPlaybackPosition
      summary: Persist a new byte position for a registered book
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPositionRequest"
      responses:
        "200":
          description: Updated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"

  /api/v1/events:
    get:
      tags: [Events]
      operationId: streamEvents
      summary: Stream playback events with Server-Sent Events
      security:
        - ApiTokenHeader: []
        - ApiTokenQuery: []
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                examples:
                  - "id: 42\nevent: progress.updated\ndata: {\"seq\":42,\"type\":\"progress.updated\"}\n\n"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

components:
  securitySchemes:
    ApiTokenHeader:
      type: apiKey
      in: header
      name: X-TTS-Token
      description: Per-run local API token printed by `tts-reader serve`.
    ApiTokenQuery:
      type: apiKey
      in: query
      name: token
      description: Query token used by the built-in browser dashboard and SSE connections.

  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    Voice:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          examples: ["Microsoft Irina Desktop"]

    VoicesResponse:
      type: object
      additionalProperties: false
      required: [voices]
      properties:
        voices:
          type: array
          items:
            $ref: "#/components/schemas/Voice"

    AddBookRequest:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path:
          type: string
          description: Absolute or relative local path to a UTF-8 text file.
          examples: ["C:\\Books\\novel.txt"]
        title:
          type: string
          examples: ["Novel"]

    Book:
      type: object
      additionalProperties: false
      description: Public book metadata. Filesystem paths and progress files are internal and are never returned by the API.
      required: [id, title, size]
      properties:
        id:
          type: string
          examples: ["book-1"]
        title:
          type: string
        size:
          type: integer
          format: int64
          minimum: 0

    BooksResponse:
      type: object
      additionalProperties: false
      required: [books]
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"

    StartPlaybackRequest:
      type: object
      additionalProperties: false
      required: [book_id]
      properties:
        book_id:
          type: string
          examples: ["book-1"]
        voice:
          type: string
          examples: ["Microsoft Irina Desktop"]
        chunk_size:
          type: integer
          minimum: 1
          maximum: 10000
          default: 250

    SetPositionRequest:
      type: object
      additionalProperties: false
      required: [book_id, current_byte]
      properties:
        book_id:
          type: string
          examples: ["book-1"]
        current_byte:
          type: integer
          format: int64
          minimum: 0

    PlaybackState:
      type: object
      additionalProperties: false
      required: [state, progress_percent, current_byte]
      properties:
        state:
          type: string
          enum: [stopped, stopping, playing, paused, finished, failed]
        book_id:
          type: string
          examples: ["book-1"]
        progress_percent:
          type: number
          format: double
          minimum: 0
          maximum: 100
        current_byte:
          type: integer
          format: int64
          minimum: 0
        voice:
          type: string
        chunk_size:
          type: integer
          minimum: 1
        error_code:
          type: string
          enum: [internal_error, playback_stopping]
        error:
          type: string

    PlaybackEvent:
      type: object
      additionalProperties: false
      required: [seq, type, time, playback]
      properties:
        seq:
          type: integer
          format: int64
          minimum: 1
        type:
          type: string
          enum:
            - playback.snapshot
            - playback.started
            - chunk.started
            - progress.updated
            - playback.paused
            - playback.resumed
            - playback.stopping
            - playback.stopped
            - playback.finished
            - playback.failed
            - position.updated
        time:
          type: string
          format: date-time
        playback:
          $ref: "#/components/schemas/PlaybackState"

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, error]
      properties:
        code:
          type: string
          examples: ["book_modified"]
        error:
          type: string
        playback:
          $ref: "#/components/schemas/PlaybackState"